moved helics into its own class
