+++ amr-wind/CFDSim.cpp
@@ -7,59 +7,7 @@
 
 #include "AMReX_ParmParse.H"
 
-
-#include "helics/cpp98/CombinationFederate.hpp"
-#include "helics/cpp98/helics.hpp"
-#include "helics/cpp98/Federate.hpp"
-
-
-#include <sstream>
-#include <iostream>
-#include <string>
-
-using namespace helicscpp;
-
 namespace amr_wind {
-
-helics_storage::helics_storage()
-	: m_fi(new helicscpp::FederateInfo("zmq"))
-    , m_vfed(new helicscpp::CombinationFederate("Test receiver Federate",*m_fi))
-{
-
-    std::cout <<"PI RECEIVER: Value federate created";
-    m_vfed->setProperty(HELICS_PROPERTY_TIME_DELTA, 1.0);
-    //  Subscribe to PI SENDER's publication 
-    // auto& in1 = vFed1->registerInput<double>("");
-    std::cout <<"PI RECEIVER: Value federate created";
-    // auto& subid = cFed->registerSubscription(vtarget + "/pub", "double");
-    auto m_sub_2 = m_vfed->registerSubscription("control", "string");
-    // m_sub = m_vfed->registerSubscription("control", "string");
-    //std::cout <<"PI RECEIVER: Subscription registered";
-    std::cout <<"PI RECEIVER: Value federate created";
-    //  Register the publication 
-    auto m_pub_2 = m_vfed->registerGlobalPublication("status", "string");
-
-
-    //std::cout <<"AMRWIND RECEIVER: Publication registered\n";
-    std::cout <<"PI RECEIVER: Value federate created";
-    //  Enter initialization state 
-    m_vfed->enterInitializingMode(); // can throw helicscpp::InvalidStateTransition exception
-    std::cout <<"PI RECEIVER: Value federate created";
-    m_vfed->enterExecutingMode(); 
-    std::cout <<"PI RECEIVER: Value federate created";
-    // currenttime = m_vfed->getCurrentTime();
-    std::cout <<"Creation complete!! ";
-    // std::cout <<"\n" << m_sub.getString();
-    
-    m_inflow_wind_speed_to_amrwind = 0.0;
-    m_inflow_wind_direction_to_amrwind = 0.0;    
-    m_turbine_power_to_controller.resize(m_num_turbines, 0.0);
-    m_turbine_yaw_to_controller.resize(m_num_turbines, 0.0);  
-    m_turbine_yaw_to_amrwind.resize(m_num_turbines, 0.0);  
-        
-}
-
-helics_storage::~helics_storage() = default;
 
 CFDSim::CFDSim(amrex::AmrCore& mesh)
     : m_mesh(mesh)
@@ -68,7 +16,7 @@ CFDSim::CFDSim(amrex::AmrCore& mesh)
     , m_io_mgr(new IOManager(*this))
     , m_post_mgr(new PostProcessManager(*this))
     , m_ext_solver_mgr(new ExtSolverMgr)
-    , m_helics(new helics_storage) {}
+    , m_helics(new helics_storage(*this)) {}
 
 CFDSim::~CFDSim() = default;
