+++ amr-wind/wind_energy/actuator/Actuator.cpp
@@ -92,6 +92,23 @@ void Actuator::pre_advance_work()
     update_velocities();
     compute_forces();
     compute_source_term();
+    
+        
+	// send power and yaw from root actuator proc to io proc
+    const int ptag = 0;
+    const int ytag = 1;
+    const size_t size = 1;
+    for (auto& ac : m_actuators) {    
+        if(ac->info().is_root_proc) {
+            amrex::ParallelDescriptor::Send(&m_sim.helics().m_turbine_power_to_controller[ac->info().id], size, amrex::ParallelDescriptor::IOProcessorNumber(), ptag);
+            amrex::ParallelDescriptor::Send(&m_sim.helics().m_turbine_yaw_to_controller[ac->info().id], size, amrex::ParallelDescriptor::IOProcessorNumber(), ytag);
+        }
+        if(amrex::ParallelDescriptor::IOProcessor()){
+            amrex::ParallelDescriptor::Recv(&m_sim.helics().m_turbine_power_to_controller[ac->info().id], size, ac->info().root_proc, ptag);
+            amrex::ParallelDescriptor::Recv(&m_sim.helics().m_turbine_yaw_to_controller[ac->info().id], size, ac->info().root_proc, ytag);
+        }
+    }
+    
 }
 
 /** Set up the container for sampling velocities
@@ -182,6 +199,7 @@ void Actuator::compute_forces()
             ac->compute_forces();
         }        
     }
+
 }
 
 void Actuator::compute_source_term()
